+++ crypto/engine/eng_lib.c
@@ -168,7 +168,6 @@ int engine_cleanup_add_last(ENGINE_CLEANUP_CB *cb)
 /* The API function that performs all cleanup */
 static void engine_cleanup_cb_free(ENGINE_CLEANUP_ITEM *item)
 {
-    (*(item->cb)) ();
     OPENSSL_free(item);
 }